rename server folder to backend folder
